app/requirements.txt
flet
pymysql
mysql-connector-python
loguru
passlib
bcrypt==4.0.1
